insta_base/requirements.txt
asgiref==3.5.2
distlib==0.3.6
Django==4.1.1
django-bootstrap5==22.1
djangorestframework==3.14.0
filelock==3.8.0
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.6
sqlparse==0.4.3
virtualenv==20.16.6
